requirements.txt
asgiref==3.4.1; python_version >= "3.6"
cachetools==3.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
certifi==2021.5.30; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
charset-normalizer==2.0.4; python_full_version >= "3.6.0" and python_version >= "3.6"
dj-database-url==0.5.0
django-admin-actions==0.1.1
django-admin-rangefilter==0.8.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
django-dbbackup==3.3.0
django-environ==0.4.5
django-sql-explorer==2.1.2; python_version >= "3.6"
django==3.2; python_version >= "3.6"
google-api-core==2.8.2; python_version >= "3.6"
google-auth==1.34.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
google-cloud-secret-manager==2.7.0; python_version >= "3.6"
googleapis-common-protos==1.56.4; python_version >= "3.7"
grpc-google-iam-v1==0.12.3; python_version >= "3.6"
grpcio-status==1.39.0; python_version >= "3.6"
grpcio==1.39.0; python_version >= "3.7"
gunicorn==20.1.0; python_version >= "3.5"
idna==3.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
libcst==0.3.19; python_version >= "3.6"
mypy-extensions==0.4.3; python_version >= "3.6"
mysqlclient==2.0.3; python_version >= "3.5"
packaging==21.0; python_version >= "3.6"
proto-plus==1.19.0; python_version >= "3.6"
protobuf==3.17.3; python_version >= "3.7"
psycopg2==2.9.3; python_version >= "3.6"
pyasn1-modules==0.2.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
pytz==2021.1; python_version >= "3.6"
pyyaml==5.4.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
requests==2.26.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
rsa==4.4; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
sqlparse==0.4.1; python_version >= "3.6"
typing-extensions==3.10.0.0; python_version >= "3.6"
typing-inspect==0.7.1; python_version >= "3.6"
urllib3==1.22; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
